% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{calerrors}
\alias{calerrors}
\title{Load namespace of CalibrationErrors.jl and CalibrationErrorsDistributions.jl}
\usage{
calerrors()
}
\description{
Load namespace of CalibrationErrors.jl and CalibrationErrorsDistributions.jl
}
