% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_weights.R
\name{getOptimalWeights}
\alias{getOptimalWeights}
\title{An Estimator for "Optimal" Weight Selection}
\usage{
getOptimalWeights(W)
}
\value{
A list containing $weights, the optimally computed weights, as well as 
   $M_j, the error-covariance structure matrix.
}
\description{
This function returns the "optimal" weights for a set of
}
\examples{
getOptimalWeights(W)
}
\seealso{
[getMj()] which this function calls for the $Mj return, [generalizedRC()] which uses these weights
}
