% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install Julia dependencies of rcalibration}
\usage{
install()
}
\description{
This function installs the Julia dependencies of rcalibration.
Note that the first initialization will take more time since it
includes precompilation.
}
