% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_weights.R
\name{solveWeights}
\alias{solveWeights}
\title{A numeric estimator for Optimal Weight Selection}
\usage{
solveWeights(W, maxit = 500, epsilon = 1e-10)
}
\arguments{
\item{W}{A list of length 'k' containing matrices of error-prone proxy measurements of the covariate. Matrices should all be n (observations) x p 
(dimension of covariates).}

\item{maxit}{The maximum number of iterations to run. Defaults to 500.}

\item{epsilon}{The tolerance at which estimates are deemed to have converged (when all weights change by less than epsilon). Defaults to 1e-10.}
}
\value{
A list containing $weights, the optimally computed weights, as well as 
   $M_j, the error-covariance structure matrix.
}
\description{
This function iteratively solves the equations to estimate for the weights which minimize the MSE of the 
 linear approximation to the truth. That is, this function returns the estimated weights, when these weights
 are considered parameters in the BLUP equation.
}
\examples{
solveWeights(W)
}
\seealso{
[rcalibration::getMj()] which this function calls for the $Mj return, [rcalibration::generalizedRC()] which uses these weights
}
