% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mj.R
\name{getMj}
\alias{getMj}
\title{An Estimator for error-covariance matrix.}
\usage{
getMj(W)
}
\arguments{
\item{W}{A list of length 'k' containing matrices of error-prone proxy measurements of the covariate. Matrices should all be n (observations) x p 
(dimension of covariates).}
}
\value{
A list M_j, the error-covariance structure matrix.
}
\description{
The generalized estimator works on error models with covariance of the form cov(X) + Mj; this function estimates the residual Mj term.
}
\examples{
getMj(W)
}
\seealso{
[getOptimalWeights()] which calls this function to compute weights [generalizedRC()] which uses this if non-optimal weights are selected
}
