% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_weights.R
\name{getOptimalWeights}
\alias{getOptimalWeights}
\title{An Estimator for "Optimal" Weight Selection}
\usage{
getOptimalWeights(W, ...)
}
\arguments{
\item{W}{A list of length 'k' containing matrices of error-prone proxy measurements of the covariate. Matrices should all be n (observations) x p 
(dimension of covariates).}
}
\value{
A list containing $weights, the optimally computed weights, as well as 
   $M_j, the error-covariance structure matrix.
}
\description{
This function returns the "optimal" weights for a set of unbiased proxies for the truth. The optimality condition is based on
 the minimization of the MSE, using the weighted proxy as an estimator for the true covariate.
}
\examples{
getOptimalWeights(W)
}
\seealso{
[rcalibration::getMj()] which this function calls for the $Mj return, [rcalibration::generalizedRC()] which uses these weights
}
